refactor: make decorations reusable
